+++ stac_fastapi/opensearch/stac_fastapi/opensearch/app.py
@@ -114,7 +114,9 @@ search_extensions = [
     SortExtension(),
     TokenPaginationExtension(),
     filter_extension,
-    FreeTextExtension(),
+    FreeTextExtension(
+        conformance_classes=[FreeTextConformanceClasses.SEARCH],
+    ),
 ]
 
 
@@ -247,6 +249,9 @@ items_get_request_model = create_request_model(
         ),
         filter_extension,
         FieldsExtension(conformance_classes=[FieldsConformanceClasses.ITEMS]),
+        FreeTextExtension(
+            conformance_classes=[FreeTextConformanceClasses.ITEMS],
+        ),
     ],
     request_type="GET",
 )